Improve CoordinatesTracker register Game are now replaying with intervalTime of CoordinatesTracker and not with a calculated time with StartedTime -> no more visual shift between gaze and actions. Change coordinates of CoordinatesTracker by pourcentage in screen for be able to replay game on other screen with different definition. Correction when cancel AOI play in AOI context.

+++ gazeplay-commons/src/main/java/net/gazeplay/commons/utils/stats/CoordinatesTracker.java
@@ -6,8 +6,8 @@ import lombok.RequiredArgsConstructor;
 @Data
 @RequiredArgsConstructor
 public class CoordinatesTracker {
-    private final int xValue;
-    private final int yValue;
+    private final double xValue;
+    private final double yValue;
     private final long intervalTime;
     private final long timeStarted;
     private final String event;
